+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/Camera/XRView.cs
@@ -1,0 +1,17 @@
+// XRTODO(2019.3) Deprecate legacy code
+// XRTODO(2020.1) Remove legacy code
+#if UNITY_2019_2_OR_NEWER
+    #define USE_XR_SDK
+#endif
+
+using System;
+using System.Collections.Generic;
+using UnityEngine.Rendering;
+#if USE_XR_SDK
+using UnityEngine.Experimental.XR;
+#endif
+
+namespace UnityEngine.Experimental.Rendering.HDPipeline
+{
+
+}