+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added an exposure node to retrieve the current, inverse and previous frame exposure value.
 - Added an HD scene color node which allow to sample the scene color with mips and a toggle to remove the exposure.
 - Added ViewConstants StructuredBuffer to help with management of XR instanced views
+- Added multipass support for VR
+- Added XR SDK integration (multipass only)
 
 ### Fixed
 - Fixed HDRI sky intensity lux mode